Move JavaCSVTableSource to Java platform and rename to JavaCSVFileSource
